+++ src/mirakuru/http.py
@@ -19,7 +19,7 @@
 
 import re
 import socket
-from urllib.parse import urlparse
+from urllib.parse import urlparse, urlencode
 from http.client import HTTPConnection, HTTPException
 
 from mirakuru.tcp import TCPExecutor
@@ -31,7 +31,10 @@ class HTTPExecutor(TCPExecutor):
     DEFAULT_PORT = 80
     """Default TCP port for the HTTP protocol."""
 
-    def __init__(self, command, url, status=r'^2\d\d$', **kwargs):
+    def __init__(
+            self, command, url, status=r'^2\d\d$', method='HEAD', payload=None,
+            **kwargs
+    ):
         """
         Initialize HTTPExecutor executor.
 
@@ -44,6 +47,9 @@ class HTTPExecutor(TCPExecutor):
             is interpreted as a single status code - e.g. '200' or '404' but
             also it can be a regular expression - e.g. '4..' or '(200|404)'.
             Default: any 2XX HTTP status code.
+        :param str method: request method to check status on.
+            Defaults to HEAD.
+        :param dict payload: Payload to send along the request
         :param int timeout: number of seconds to wait for the process to start
             or stop. If None or False, wait indefinitely.
         :param float sleep: how often to check for start/stop condition
@@ -66,6 +72,8 @@ class HTTPExecutor(TCPExecutor):
 
         self.status = str(status)
         self.status_re = re.compile(str(status))
+        self.method = method
+        self.payload = payload
 
         super(HTTPExecutor, self).__init__(
             command, host=self.url.hostname, port=port, **kwargs
@@ -75,8 +83,14 @@ class HTTPExecutor(TCPExecutor):
         """Check if defined URL returns expected status to a HEAD request."""
         try:
             conn = HTTPConnection(self.host, self.port)
-
-            conn.request('HEAD', self.url.path)
+            if self.payload and self.method in ('POST', 'PUT', 'PATCH'):
+                conn.request(
+                    self.method,
+                    self.url.path,
+                    urlencode(self.payload)
+                )
+            else:
+                conn.request(self.method, self.url.path)
             status = str(conn.getresponse().status)
 
             if status == self.status or self.status_re.match(status):